+++ 9465_Stefano_tut3_task3_4/src/StefanoTut3/Tut3Task4Sphere.java
@@ -10,7 +10,7 @@ public class Tut3Task4Sphere extends Tutorial3Task3
     public void calculateArea() 
     {
         String curve = JOptionPane.showInputDialog(new JFrame(),"Enter the curve of the Sphere: ");
-        int sphere = (int) (3.14 * (Integer.valueOf(radius) * Integer.valueOf(radius))) + Integer.valueOf(curve);
+        Double sphere = (3.14 * radius * radius) + Integer.valueOf(curve);
         JOptionPane.showMessageDialog(new JFrame(), "The Sphere is : " + sphere); 
     }    
 }
